+++ env.js
@@ -1,0 +1,1 @@
+self.DEV_MODE = true;